+++ custom_components/house_voice/const.py
@@ -1,9 +1,9 @@
-# VERSION = "2.2.0"
+# VERSION = "3.0.0"
 # File: const.py
 # Description: Constants for House Voice Manager
 
 DOMAIN   = "house_voice"
-VERSION  = "2.2.0"
+VERSION  = "3.0.0"
 
 # Storage
 STORAGE_KEY        = "house_voice_events"